* 'CRM_Release_1.50.0.0' of bitbucket.org:anuj_aggarwal/247around-adminp-aws:
  #CRM-469 add Rm email id in the upcountry failed mail
  #CRM-468
  #CRM-465  change amazon file format
  #CRM-465 change amazon file format
  Do not send customer invoice sms when pdf not generated
  Add seller invoice ID in view details Page #CRM-470
  Set Content type in Qr Image

+++ application/controllers/buyback/buyback_process.php
@@ -1368,34 +1368,63 @@ class Buyback_process extends CI_Controller {
         return $flag;
     }
     
+    
+     /**
+     * @desc This function is used to download the amazon price list
+     * @param void
+     * @return void
+     */
     public function download_price_list_data() {
-
-        $csv_file_name = TMP_FOLDER . "buyback_price_list";
-        $where = array('bb_charges.partner_id' => AMAZON_SELLER_ID, 'visible_to_partner' => 1, 'bb_shop_address.active' => 1);
-        $total_data = $this->service_centre_charges_model->get_bb_charges($where, 'count(bb_charges.id) as total_data', true, true);
-        $row_limit = 500;
-        $counter = round($total_data[0]['total_data'] / $row_limit);
-        $offset = 0;
-        for ($i = 0; $i < $counter; $i++) {
-            $select = "category,brand, physical_condition, working_condition , city AS location , partner_total";
-            $data = $this->service_centre_charges_model->get_bb_charges($where, $select, true, true, $offset, $row_limit, TRUE);
-            if (!empty($data)) {
-                $file_name = $csv_file_name . "_" . $i;
-                $csv_file[$file_name] = $this->generate_bb_csv_price_list($file_name, $data);
+        $service_name_arr = $this->input->post('appliance_name');
+        if(!empty($service_name_arr)) {
+            
+            //If all is selected then download all appliance data
+            $key = array_search('All', $service_name_arr);
+            if ($key !== FALSE) {
+                $service_name_arr = array_column($this->booking_model->selectservice(true),'services', 'id');
             }
-            $offset += 500;
-        }
-        
-        $this->load->library('zip');
-        
-        foreach (array_keys($csv_file) as $value) {
-            $this->zip->read_file($value);
-            $res1 = 0;
-            system("chmod 777" . $value, $res1);
-            unlink($value);
-        }
+            
+            foreach ($service_name_arr as $key => $value) {
+                $csv_file_name = TMP_FOLDER . "buyback_price_list_" . strtolower(str_replace(" ", "_", $value));
+                $where = array('bb_charges.partner_id' => AMAZON_SELLER_ID, 'visible_to_partner' => 1, 'bb_shop_address.active' => 1, 'bb_charges.service_id' => $key);
+                //get total data and divide it from 500 to insert only 500 data at a time in one csv
+                $total_data = $this->service_centre_charges_model->get_bb_charges($where, 'count(bb_charges.id) as total_data', true, true);
+                if (!empty($total_data)) {
+                    $row_limit = 500;
+                    $counter = round($total_data[0]['total_data'] / $row_limit);
+                    $offset = 0;
+                    for ($i = 0; $i < $counter; $i++) {
+                        if($key == _247AROUND_MOBILE_SERVICE_ID){
+                            $select ="brand as Brand, physical_condition as 'Model' , city AS 'Location' , partner_total as 'Exchange Offer Value'";
+                        }else{
+                            $select = "brand as Brand,category as Type, concat(physical_condition, ' | ',working_condition) as 'Product Condition' , city AS 'Location' , partner_total as 'Exchange Offer Value'";
+                        }
+                        $data = $this->service_centre_charges_model->get_bb_charges($where, $select, true, true, $offset, $row_limit, TRUE);
+                        if (!empty($data)) {
+                            $file_name = $csv_file_name . "_" . $i;
+                            $csv_file[$file_name] = $this->generate_bb_csv_price_list($file_name, $data);
+                        }
+                        $offset += 500;
+                    }
+                }
+            }
 
-        $this->zip->download('buyback_price_list.zip');
+            if(!empty($csv_file)){
+                //zipped all the files and download it
+                $this->load->library('zip');
+
+                foreach (array_keys($csv_file) as $value) {
+                    $this->zip->read_file($value);
+                    $res1 = 0;
+                    system("chmod 777" . $value, $res1);
+                    unlink($value);
+                }
+
+                $this->zip->download('buyback_price_list.zip');
+            }
+        }else{
+            echo "Empty data submitted";
+        }
     }
 
     /**
